% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ccs.test}
\alias{ccs.test}
\title{Estimates a common cointegration space as in Callot (2014).}
\usage{
ccs.test(Y, W = NULL, rank = 1, lags = 2, BS = 99, ddet = 0,
  Y0 = NULL, seasonal = FALSE, ccsinit = NULL, cores = 1,
  bs.method = "resample")
}
\arguments{
\item{Y}{A molten data set containing the 5 columns. *indiv*: the individual (country, sector,...) name as factor levels. *Year*: the year of observation. *Period*: the period (day, week, month, quarter) index as consecutive integers. *Variable*: the name of the variable. *value*: the value of the variable at the given time for a given individual.}

\item{W}{either NULL, A matrix, 3 dimensional array with first and second dimension equal to the number of individuals. Columns must sum to one, and values on the diagonal must be zero. In case the input is an array, its third dimension should be equal to the number of observations. In case of NULL, equal weights are used.}

\item{rank}{An integer greater or equal to 1 indicating the co-integration rank of the individual systems.}

\item{lags}{An integer indicating the number of lags of the VAR in level.}

\item{BS}{The number of bootstrap iteration. Default: 99.}

\item{ddet}{An integer indicating the type of deterministics to use, following the typology by Johansen 1988.}

\item{Y0}{A molten set of global exogenous data to use. Its columns must be identical to those of Y, and have the same number of observations.}

\item{seasonal}{A boolean indicating whether seasonal dummies should be used. Default FALSE}

\item{ccsinit}{Optional, a matrix with *rank* columns containing the inital value. If NULL, default to the average of the unconstrained estimator.}

\item{cores}{An integer greater or equal to 1. if it is greater than 1 the bootstrap is parallelized with mclapply. default to 1.}

\item{bs.method}{The type of bootstrap to use: _resample_ or _wild_.}
}
\value{
A list with *stat* the panel trace test statistic for each hypothesis. *pval* the corresponding p-value. A number or other useful statistics are returned.
}
\description{
bla!
}
\details{
bla, bla?
}
\author{
Laurent Callot \email{l.callot@vu.nl}
}

