% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ll.test}
\alias{ll.test}
\title{Computes the common cointegration space estimator of Larsson and Lyhagen 2007 JBES}
\usage{
ll.test(Y, betainit, rank = 1, lags = 2, ddet = 1, Y0 = NULL,
  seasonal = FALSE)
}
\arguments{
\item{Y}{A molten data set containing the 4 columns. indiv: the individual (country, sector,...) name as factor levels. Year: the year of observation. Period: the period (day, week, month, quarter) index as consecutive integers. Variable: the name of the variable. value: the value of the variable at the given time for a given individual.}

\item{betainit}{Initial value for the co-integration vectors.}

\item{rank}{The co-integration rank of the system, default 1.}

\item{lags}{An integer indicating the number of lags of the VAR in level.}

\item{ddet}{An integer indicating the type of deterministics to use, following the typology by Johansen 1988.}

\item{Y0}{Optional: a molten set of global exogenous data to use. Its columns must be identical to those of Y, and have the same number of observations.}

\item{seasonal}{Optional, should seasonal dummies be included. Default FALSE.}
}
\value{
A list with 'est' the individual CVAR and 'trace' the trace test statistics for each model and each hypothesis.
}
\description{
bla!
}
\details{
bla, bla?
}
\author{
Laurent Callot \email{l.callot@vu.nl}
}

