% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mc.ccs.test}
\alias{mc.ccs.test}
\title{Performs a Monte Carlo experiment on the performance of the panel Common Co-integration space estimator and test of Callot (2014).}
\usage{
mc.ccs.test(obs, N, nvar, rank = 1, BS = 99, MC = 1, W = "equal",
  Alpha = NULL, Beta = NULL, Lambda0 = NULL, Gammal = NULL, Omega,
  err.dist = "gaussian", det.type = 1, cdet.load = c(1, 1),
  bs.method = "resample", t.df = 3, burn.smpl = 10, res.dep = "iid",
  garchspec = NULL, ncores = 1, llest = TRUE)
}
\arguments{
\item{obs}{An integer, the number of observations to generate.}

\item{N}{an integer, the number of cross section units.}

\item{nvar}{The number of variables for each cross section unit.}

\item{rank}{The cointegration rank of the system. Default 1.}

\item{BS}{The number of bootstrap iterations for the test. Default: 99.}

\item{MC}{The number of monte carlo iterations. Default: 1.}

\item{Alpha}{Adjustment matrix (dimension nvar * rank) or list (length N) of parameter matrices.}

\item{Beta}{Co-integration matrix (dimension 2nvar * rank) or list (length N) of parameter matrices.}

\item{Lambda0}{Contemporaneous dependency matrix (dimension nvar  *  nvar) or list (length N) of parameter matrices.}

\item{Gammal}{List (length is number of lagged first differences) of matrices (N  *  N) or lists (length N) of matrices.}

\item{Omega}{The covariance matrix.}

\item{err.dist}{The distribution of the innovations, _gaussian_ or _t_}

\item{t.df}{If the innovations are _t_ distributed, the number of degrees of freedom. default 3.}

\item{det.type}{An integer indicating the type of deterministics to use, following the typology by Johansen 1988.}

\item{burn.smpl}{The Number of burned observations used to generate the data.}

\item{res.dep}{Dependency of the residuals, _iid_ (default) or _garch_.}

\item{garchspec}{See fGarch package.}

\item{cdet.load}{The loadings on the deterministics.}

\item{W}{The weighting scheme. Default: equal.}

\item{bs.method}{'resample' for iid resampling, 'wild' for gaussian wild bootstrap.}

\item{ncores}{The number of cores, default 1.}

\item{llest}{Should the Larsson and Lyhagen 2007 JBES CCS estimator be computed? Default FALSE.}
}
\value{
A list.
}
\description{
bla!
}
\details{
bla, bla?
}
\author{
Laurent Callot \email{l.callot@vu.nl}
}

