% Generated by roxygen2 (4.0.1): do not edit by hand
\name{meltdat}
\alias{meltdat}
\title{Transforms a matrix of data to a molten data frame for use in rank.test..}
\usage{
meltdat(Ymat, indiv.names, var.names, time.index = 1:nrow(Ymat), freq = 1)
}
\arguments{
\item{Ymat}{A matrix of data with variables in columns. The variables are assumed grouped by countries.}

\item{indiv.names}{The names of the individuals}

\item{var.names}{The names of the variables}

\item{time.index}{A time index, default to 1:T.}

\item{freq}{The frequency of the seasons, default 1 if not seasonal.}
}
\value{
A molten dataframe ready for use in rank test.
}
\description{
bla!
}
\details{
bla, bla?
}
\author{
Laurent Callot \email{l.callot@vu.nl}
}

