% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gen.pcvar}
\alias{gen.pcvar}
\title{Generates data from a PCVAR model.}
\usage{
gen.pcvar(obs, N, nvar, W = NULL, Alpha = NULL, Beta = NULL, Lambda0,
  Gammal = NULL, Omega, err.dist = "gaussian", burn.smpl = 10, t.df = 1,
  det.type = 1, cdet.load = c(1, 1), res.dep, garchspec)
}
\arguments{
\item{obs}{An integer, the number of observations to generate.}

\item{N}{an integer, the number of cross section units.}

\item{nvar}{The number of variables for each cross section unit.}

\item{Alpha}{Adjustment matrix (dimension nvar * rank) or list (length N) of parameter matrices.}

\item{Beta}{Co-integration matrix (dimension 2nvar * rank) or list (length N) of parameter matrices.}

\item{Lambda0}{Contemporaneous dependency matrix (dimension nvar  *  nvar) or list (length N) of parameter matrices.}

\item{Gammal}{List (length is number of lagged first differences) of matrices (N  *  N) or lists (length N) of matrices.}

\item{Omega}{The covariance matrix.}

\item{err.dist}{The distribution of the innovations, _gaussian_ or _t_}

\item{t.df}{If the innovations are _t_ distributed, the number of degrees of freedom. default 3.}

\item{det.type}{An integer indicating the type of deterministics to use, following the typology by Johansen 1988.}

\item{burn.smpl}{The Number of burned observations used to generate the data.}

\item{res.dep}{Dependency of the residuals, _iid_ (default) or _garch_.}

\item{garchspec}{See fGarch package.}

\item{cdet.load}{The loadings on the deterministics.}

\item{W}{The weighting scheme. Default: equal.}
}
\value{
A list.
}
\description{
bla!
}
\details{
bla, bla?
}
\author{
Laurent Callot \email{l.callot@vu.nl}
}

